🛠️ 🏗️ feat: implement streamed translation renderer with batched block reveal and smooth modal height growth
